Allow for missing minor and patch in version string
To deal with version strings that don’t include a minor and/or patch, we also needed to update the way version strings are created.
It is now possible to pass some formatting options to the versionString function to control what is included in the string and what isn’t.